+++ dscreate/create/SplitNotebook.py
@@ -14,26 +14,28 @@ class SplitNotebook:
             is added to the solution notebook and is replaced with
             `*YOUR ANSWER HERE*` in the lesson notebook.
     """
-    def __init__(self):
-        self.solution_tags = ["__SOLUTION__", f"#__SOLUTION__", 
+    def __init__(self, filename='index.ipynb'):
+        solution_tags = ["__SOLUTION__", f"#__SOLUTION__", 
                             "==SOLUTION==", f"#==SOLUTION=="]
-        self.solution_cells = []
-        self.lesson_cells = []
+        self.data = self.get_notebook_json(filename=filename)
 
-    def get_notebook_json(self, filename="curriculum.ipynb"):
+    def get_notebook_json(self, filename="index.ipynb"):
         """
         Import raw json data for a source notebook
         """
         with open(filename, 'r') as f:
-            self.data = json.load(f)
+            data = json.load(f)
+        return data
         
-    def split_notebooks(self):
+    def split_notebook(self):
         """
         Loops over each cell in a notebook
         and appends all solution and lesson cells to the
         `solution_cells` and `lesson_cells` attributes.
         """
         count = 0
+        lesson_cells = []
+        solution_cells = []
         for cell in self.data['cells']:
             cell_type = cell['cell_type']
             solution, lines = self._parse_cell(cell)
@@ -45,20 +47,21 @@ class SplitNotebook:
             if cell_type == "markdown":
 
                 if solution:
-                    self.solution_cells.append(cell)
+                    solution_cells.append(cell)
                     placeholder.update({"source": ['*YOUR ANSWER HERE*']})
-                    self.lesson_cells.append(placeholder)
+                    lesson_cells.append(placeholder)
                 else:
-                    self.lesson_cells.append(cell)
-                    self.solution_cells.append(placeholder)
+                    lesson_cells.append(cell)
+                    solution_cells.append(placeholder)
             else:
                 if solution:
-                    self.solution_cells.append(cell)
+                    solution_cells.append(cell)
 
                 else:
-                    self.lesson_cells.append(cell)
-                    self.solution_cells.append(placeholder) 
-            count += 1  
+                    lesson_cells.append(cell)
+                    solution_cells.append(placeholder) 
+            count += 1 
+        return lesson_cells, solution_cells 
 
     def _parse_cell(self, cell):
         """
@@ -73,7 +76,7 @@ class SplitNotebook:
         lines = []
         for line in cell["source"]:
             found = False
-            for tag in self.solution_tags:
+            for tag in solution_tags:
                 if tag in line.strip().split(" "):
                     is_solution = True
                     found = True
@@ -93,20 +96,3 @@ class SplitNotebook:
         file = open(f"{name}.ipynb", "w")
         file.write(json.dumps(notebook))
         file.close()
-
-    def main(self, filename="curriculum.ipynb"):
-        """
-        Main function for running the splitting process.
-        """
-        self.get_notebook_json(filename=filename)
-        self.split_notebooks()
-        solution_dir = '.solution_files'
-        if not os.path.isdir(solution_dir):
-            os.mkdir(solution_dir)
-        solution_path = os.path.join(solution_dir, 'index')
-        self.write_notebook(solution_path, self.solution_cells)
-        self.write_notebook('index', self.lesson_cells)
-
-if __name__ == "__main__":
-    splitter = SplitNotebook()
-    splitter.main()
