+++ dscreate/create/__init__.py
@@ -1,5 +1,6 @@
 from .ShareNotebook import ShareNotebook
 from .SplitNotebook import SplitNotebook
 from .Color import Color
-from .BeginLesson import begin
+from .begin import begin
 from .GenerateCurriculum import GenerateCurriculum
+from .readme import generate_readme
