+++ dscreate/utils/__init__.py
@@ -1,1 +1,3 @@
 from .setup import spark_setup
+from .Git import *
+from .helpers import *
