+++ setup.py
@@ -1,7 +1,7 @@
 from setuptools import setup, find_packages
 
 setup(name='dscreate',
-      version='0.1.84',
+      version='0.1.85',
       description='Flatiron Iron School Data Science Tools',
       url='http://github.com/learn-co-curriculum/dscreate',
       author='Joél Collins',
